requirements.txt
asgiref==3.3.4
Django==3.2.3
djangorestframework==3.12.4
Pygments==2.9.0
pytz==2021.1
sqlparse==0.4.1
